+++ geotrek/tourism/management/commands/sync_rando.py
@@ -6,9 +6,11 @@ from django.conf import settings
 from django.utils import translation, timezone
 
 from geotrek.tourism import models as tourism_models
-from geotrek.tourism.views import (TouristicContentViewSet, TouristicEventViewSet,
-                                   TrekTouristicContentViewSet, TrekTouristicEventViewSet)
+from geotrek.tourism import views as tourism_views
 from geotrek.trekking.management.commands.sync_rando import Command as BaseCommand
+
+# Register mapentity models
+from geotrek.trekking import urls  # NOQA
 
 
 class Command(BaseCommand):
@@ -49,10 +51,11 @@ class Command(BaseCommand):
             self.sync_media_file(lang, resized)
 
     def sync_tourism(self, lang):
-        self.sync_geojson(lang, TouristicContentViewSet, 'touristiccontents')
-        self.sync_geojson(lang, TouristicEventViewSet, 'touristicevents', params={'ends_after': timezone.now().strftime('%Y-%m-%d')})
+        self.sync_geojson(lang, tourism_views.TouristicContentViewSet, 'touristiccontents')
+        self.sync_geojson(lang, tourism_views.TouristicEventViewSet, 'touristicevents',
+                          params={'ends_after': timezone.now().strftime('%Y-%m-%d')})
 
-        # picto in global zip
+        # reopen global zip (closed after trekking sync)
         self.zipfile = ZipFile(os.path.join(self.tmp_root, 'zip', 'treks',
                                             lang, 'global.zip'),
                                'a')
@@ -63,6 +66,19 @@ class Command(BaseCommand):
                        settings.STATIC_ROOT,
                        settings.STATIC_URL,
                        zipfile=self.zipfile)
+
+        # json with
+        params = {}
+
+        if self.categories:
+            params.update({'categories': ','.join(category for category in self.categories), })
+
+        if self.with_events:
+            params.update({'events': '1'})
+
+        self.sync_json(lang, tourism_views.TouristicCategoryView,
+                       'touristiccategories',
+                       zipfile=self.zipfile, params=params)
 
         # pictos touristic content catgories
         for category in tourism_models.TouristicContentCategory.objects.all():
@@ -89,7 +105,7 @@ class Command(BaseCommand):
         params = {'format': 'geojson',
                   'categories': ','.join(category for category in self.categories), }
 
-        view = TrekTouristicContentViewSet.as_view({'get': 'list'})
+        view = tourism_views.TrekTouristicContentViewSet.as_view({'get': 'list'})
         name = os.path.join('api', lang, 'treks', str(trek.pk), 'touristiccontents.geojson')
         self.sync_view(lang, view, name, params=params, zipfile=zipfile, pk=trek.pk)
 
@@ -98,7 +114,7 @@ class Command(BaseCommand):
 
     def sync_trek_touristicevents(self, lang, trek, zipfile=None):
         params = {'format': 'geojson', }
-        view = TrekTouristicEventViewSet.as_view({'get': 'list'})
+        view = tourism_views.TrekTouristicEventViewSet.as_view({'get': 'list'})
         name = os.path.join('api', lang, 'treks', str(trek.pk), 'touristicevents.geojson')
         self.sync_view(lang, view, name, params=params, zipfile=zipfile, pk=trek.pk)
 
@@ -126,7 +142,7 @@ class Command(BaseCommand):
         self.sync_pictograms('**', tourism_models.TouristicContentType)
         self.sync_pictograms('**', tourism_models.TouristicEventType)
 
-        for lang in settings.MODELTRANSLATION_LANGUAGES:
+        for lang in self.languages:
             translation.activate(lang)
             self.sync_tourism(lang)