+++ geotrek/tourism/views.py
@@ -1,5 +1,7 @@
 from itertools import chain
 import logging
+import os
+from django.utils.translation import ugettext as _
 
 from django.conf import settings
 from django.contrib.auth.decorators import login_required
@@ -17,12 +19,16 @@ from mapentity.settings import app_settings as mapentity_settings
 import requests
 from requests.exceptions import RequestException
 from rest_framework import permissions as rest_permissions, viewsets
-from rest_framework_gis.serializers import GeoFeatureModelSerializer
 from rest_framework.filters import DjangoFilterBackend
+from rest_framework.renderers import JSONRenderer
+from rest_framework.response import Response
+from rest_framework.views import APIView
+from rest_framework_gis.serializers import GeoFeatureModelSerializer
 
 from geotrek.authent.decorators import same_structure_required
 from geotrek.common.utils import plain_text_preserve_linebreaks
 from geotrek.common.views import DocumentPublic
+from geotrek.tourism.serializers import TouristicContentCategorySerializer
 from geotrek.trekking.models import Trek
 from geotrek.trekking.serializers import POISerializer
 
@@ -33,6 +39,7 @@ from .models import (TouristicContent, TouristicEvent, TouristicContentCategory,
                      DataSource, InformationDesk)
 from .serializers import (TouristicContentSerializer, TouristicEventSerializer,
                           InformationDeskSerializer)
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
 
 
 logger = logging.getLogger(__name__)
@@ -269,6 +276,20 @@ class TouristicContentViewSet(MapEntityViewSet):
         return qs
 
 
+class TouristicContentCategoryViewSet(MapEntityViewSet):
+    model = TouristicContentCategory
+    serializer_class = TouristicContentCategorySerializer
+    permission_classes = [rest_permissions.DjangoModelPermissionsOrAnonReadOnly]
+
+    def get_queryset(self):
+        queryset = super(TouristicContentCategoryViewSet, self).get_queryset()
+
+        if 'categories' in self.request.GET:
+            queryset = queryset.filter(pk__in=self.request.GET['categories'].split(','))
+
+        return queryset
+
+
 class TouristicEventViewSet(MapEntityViewSet):
     model = TouristicEvent
     serializer_class = TouristicEventSerializer
@@ -370,3 +391,39 @@ class TrekTouristicEventViewSet(viewsets.ModelViewSet):
                                                    field_name='geom')
 
         return queryset
+
+
+class TouristicCategoryView(APIView):
+    """
+    touristiccategories.json generation for API
+    """
+    renderer_classes = (JSONRenderer,)
+    permission_classes = (IsAuthenticatedOrReadOnly,)
+
+    def get(self, request, format=None, lang=None):
+        response = []
+        content_categories = TouristicContentCategory.objects.all()
+
+        if request.GET.get('categories', False):
+            categories = request.GET['categories'].split(',')
+            content_categories.filter(pk__in=categories)
+
+        for cont_cat in content_categories:
+            response.append({'id': cont_cat.prefixed_id,
+                             'label': cont_cat.label,
+                             'type1_label': cont_cat.type1_label,
+                             'type2_label': cont_cat.type2_label,
+                             'pictogram': os.path.join(settings.MEDIA_URL, cont_cat.pictogram.url),
+                             'order': cont_cat.order,
+                             'slug': 'touristic-content'})
+
+        if request.GET.get('events', False):
+            response.append({'id': 'E',
+                             'label': _(u"Touristic events"),
+                             'type1_label': "",
+                             'type2_label': "",
+                             'pictogram': os.path.join(settings.STATIC_URL, 'tourism', 'touristicevent.svg'),
+                             'order': None,
+                             'slug': 'touristic-event'})
+
+        return Response(response)