+++ geotrek/trekking/management/commands/sync_rando.py
@@ -17,7 +17,6 @@ from django.test.client import RequestFactory
 from django.utils import translation
 
 from mapentity.settings import app_settings as mapentity_settings
-
 from geotrek.altimetry.views import ElevationProfile, ElevationArea, serve_elevation_chart
 from geotrek.common import models as common_models
 from geotrek.trekking import models as trekking_models
@@ -85,6 +84,8 @@ class Command(BaseCommand):
                     default=False, help='Skip generation of DEM files for mobile app'),
         make_option('--skip-profile-png', '-e', action='store_true', dest='skip_profile_png',
                     default=False, help='Skip generation of PNG elevation profile'),
+        make_option('--languages', '-l', action='store', dest='languages',
+                    default='', help='Languages to sync'),
     )
 
     def mkdirs(self, name):
@@ -182,6 +183,13 @@ class Command(BaseCommand):
             zipfile.write(fullname, name)
         if self.verbosity == '2':
             self.stdout.write(u"\x1b[3D\x1b[32mgenerated\x1b[0m")
+
+    def sync_json(self, lang, viewset, name, zipfile=None, params={}, **kwargs):
+        view = viewset.as_view()
+        name = os.path.join('api', lang, '{name}.json'.format(name=name))
+        if self.source:
+            params['source'] = ','.join(self.source)
+        self.sync_view(lang, view, name, params=params, zipfile=zipfile, **kwargs)
 
     def sync_geojson(self, lang, viewset, name, zipfile=None, params={}, **kwargs):
         view = viewset.as_view({'get': 'list'})
@@ -378,9 +386,17 @@ class Command(BaseCommand):
     def sync(self):
         self.sync_tiles()
 
-        for lang in settings.MODELTRANSLATION_LANGUAGES:
+        for lang in self.languages:
             translation.activate(lang)
             self.sync_trekking(lang)
+
+    def check_dst_root_is_empty(self):
+        if not os.path.exists(self.dst_root):
+            return
+        existing = set([os.path.basename(p) for p in os.listdir(self.dst_root)])
+        remaining = existing - set(('api', 'media', 'static', 'zip'))
+        if remaining:
+            raise CommandError(u"Destination directory contains extra data")
 
     def handle(self, *args, **options):
         self.successfull = True
@@ -388,11 +404,7 @@ class Command(BaseCommand):
         if len(args) < 1:
             raise CommandError(u"Missing parameter destination directory")
         self.dst_root = args[0].rstrip('/')
-        if os.path.exists(self.dst_root):
-            existing = set([os.path.basename(p) for p in os.listdir(self.dst_root)])
-            remaining = existing - set(('api', 'media', 'static', 'zip'))
-            if remaining:
-                raise CommandError(u"Destination directory contains extra data")
+        self.check_dst_root_is_empty()
         if(options['url'][:7] != 'http://'):
             raise CommandError('url parameter should start with http://')
         self.referer = options['url']
@@ -404,6 +416,10 @@ class Command(BaseCommand):
         self.skip_dem = options['skip_dem']
         self.skip_profile_png = options['skip_profile_png']
         self.source = options['source']
+        if options['languages']:
+            self.languages = options['languages'].split(',')
+        else:
+            self.languages = settings.MODELTRANSLATION_LANGUAGES
         if self.source is not None:
             self.source = self.source.split(',')
         self.builder_args = {